+++ Sources/Core/SnabblePay.swift
@@ -1,0 +1,12 @@
+//
+//  SnabblePay.swift
+//  
+//
+//  Created by Andreas Osberghaus on 2022-12-20.
+//
+
+import Foundation
+
+struct SnabblePay {
+    
+}