+++ Sources/Network/Core/NetworkManager.swift
@@ -1,0 +1,52 @@
+//
+//  NetworkManager.swift
+//  
+//
+//  Created by Andreas Osberghaus on 2022-12-19.
+//
+
+import Combine
+import Foundation
+
+public struct NetworkManager {
+    public let session: URLSession
+    let authenticator: Authenticator
+
+    public static let decoder: JSONDecoder = {
+        let decoder = JSONDecoder()
+        decoder.dateDecodingStrategy = .iso8601
+        return decoder
+    }()
+
+    public init(session: URLSession = .shared) {
+        self.session = session
+        self.authenticator = Authenticator(session: session)
+    }
+
+    public func publisher<Response: Decodable>(for endpoint: Endpoint<Response>) -> AnyPublisher<Response, Swift.Error> {
+        return authenticator.validToken(using: Self.decoder, onEnvironment: endpoint.environment)
+            .map { token in
+                var endpoint = endpoint
+                endpoint.token = token
+                return endpoint
+            }
+            .flatMap { endpoint in
+                session.publisher(for: endpoint, using: Self.decoder)
+            }
+            .tryCatch { error in
+                if case HTTPError.invalidResponse(let statusCode) = error, statusCode == .unauthorized {
+                    return authenticator.validToken(using: Self.decoder, forceRefresh: true, onEnvironment: endpoint.environment)
+                        .map { token in
+                            var endpoint = endpoint
+                            endpoint.token = token
+                            return endpoint
+                        }
+                        .flatMap { endpoint in
+                            session.publisher(for: endpoint, using: Self.decoder)
+                        }
+                }
+                throw error
+            }
+            .eraseToAnyPublisher()
+    }
+}