+++ Tests/Network/Core/URLSessionEndpointTests.swift
@@ -7,13 +7,13 @@
 
 import XCTest
 import Combine
-@testable import SnabblePayCore
+@testable import SnabblePayNetwork
 
 final class URLSessionEndpointTests: XCTestCase {
 
-    let resourceData = try! loadResource(filename: "credentials", withExtension: "json")
-    let endpointCredentials: Endpoint<Credentials> = .credentials()
-    let endpointData: Endpoint<Data> = .init(path: "/apps/credentials", method: .get(nil))
+    let resourceData = try! loadResource(filename: "register", withExtension: "json")
+    let endpointRegister: Endpoint<App> = Endpoints.register()
+    let endpointData: Endpoint<Data> = .init(path: "/apps/register", method: .get(nil))
     var cancellables = Set<AnyCancellable>()
 
     // MARK: - Decodable
@@ -22,7 +22,7 @@ final class URLSessionEndpointTests: XCTestCase {
         MockURLProtocol.error = nil
         MockURLProtocol.requestHandler = { [unowned self] request in
             let response = HTTPURLResponse(
-                url: URL(string: "https://payment.snabble.io/apps/credentials")!,
+                url: URL(string: "https://payment.snabble.io/apps/register")!,
                 statusCode: 200,
                 httpVersion: nil,
                 headerFields: ["Content-Type": "application/json"]
@@ -30,9 +30,9 @@ final class URLSessionEndpointTests: XCTestCase {
             return (response, resourceData)
         }
 
-        let expectation = expectation(description: "credentials")
+        let expectation = expectation(description: "register")
         let session = URLSession.mockSession
-        session.publisher(for: endpointCredentials)
+        session.publisher(for: endpointRegister)
             .sink { completion in
                 switch completion {
                 case .finished:
@@ -53,17 +53,48 @@ final class URLSessionEndpointTests: XCTestCase {
         MockURLProtocol.error = URLError(.unknown)
         MockURLProtocol.requestHandler = nil
 
-        let expectation = expectation(description: "credentials")
+        let expectation = expectation(description: "register")
         let session = URLSession.mockSession
-        session.publisher(for: endpointCredentials)
+        session.publisher(for: endpointRegister)
             .sink { completion in
                 switch completion {
                 case .finished:
                     XCTAssertTrue(true)
                 case .failure(let error):
                     XCTAssertNotNil(error)
-                    XCTAssertNotNil(error as? URLError)
-                    XCTAssertEqual((error as! URLError).code, .unknown)
+                    XCTAssertNotNil(error is URLError)
+                }
+                expectation.fulfill()
+            } receiveValue: { credentials in
+                XCTAssertNil(credentials)
+            }
+            .store(in: &cancellables)
+
+        wait(for: [expectation], timeout: 5.0)
+    }
+
+    func testDecodableCombineInvalidResponse() async throws {
+        MockURLProtocol.error = nil
+        MockURLProtocol.requestHandler = { [unowned self] request in
+            let response = HTTPURLResponse(
+                url: URL(string: "https://payment.snabble.io/apps/register")!,
+                statusCode: 400,
+                httpVersion: nil,
+                headerFields: ["Content-Type": "application/json"]
+            )!
+            return (response, resourceData)
+        }
+
+        let expectation = expectation(description: "register")
+        let session = URLSession.mockSession
+        session.publisher(for: endpointRegister)
+            .sink { completion in
+                switch completion {
+                case .finished:
+                    XCTAssertTrue(true)
+                case .failure(let error):
+                    XCTAssertNotNil(error)
+                    XCTAssertTrue(error is HTTPError)
                 }
                 expectation.fulfill()
             } receiveValue: { credentials in
@@ -78,7 +109,7 @@ final class URLSessionEndpointTests: XCTestCase {
         MockURLProtocol.error = nil
         MockURLProtocol.requestHandler = { [unowned self] request in
             let response = HTTPURLResponse(
-                url: URL(string: "https://payment.snabble.io/apps/credentials")!,
+                url: URL(string: "https://payment.snabble.io/apps/register")!,
                 statusCode: 200,
                 httpVersion: nil,
                 headerFields: ["Content-Type": "application/json"]
@@ -87,7 +118,7 @@ final class URLSessionEndpointTests: XCTestCase {
         }
 
         let session = URLSession.mockSession
-        let decodableObject = try await session.object(for: endpointCredentials)
+        let decodableObject = try await session.object(for: endpointRegister)
 
         XCTAssertNotNil(decodableObject)
     }
@@ -98,64 +129,34 @@ final class URLSessionEndpointTests: XCTestCase {
 
         do {
             let session = URLSession.mockSession
-            let decodableObject = try await session.object(for: endpointCredentials)
+            let decodableObject = try await session.object(for: endpointRegister)
             XCTAssertNil(decodableObject)
         } catch {
             XCTAssertNotNil(error)
-            XCTAssertNotNil(error as? URLError)
-            XCTAssertEqual((error as! URLError).code, .unknown)
+            XCTAssertNotNil(error is URLError)
         }
     }
 
-    func testDecodable() async throws {
+    func testDecodableAsyncInvalidResponse() async throws {
         MockURLProtocol.error = nil
         MockURLProtocol.requestHandler = { [unowned self] request in
             let response = HTTPURLResponse(
-                url: URL(string: "https://payment.snabble.io/apps/credentials")!,
-                statusCode: 200,
+                url: URL(string: "https://payment.snabble.io/apps/register")!,
+                statusCode: 400,
                 httpVersion: nil,
                 headerFields: ["Content-Type": "application/json"]
             )!
             return (response, resourceData)
         }
-        let expectation = expectation(description: "credentials")
-        let session = URLSession.mockSession
-        let dataTask = session.dataTask(for: endpointCredentials) { result in
-            switch result {
-            case .failure(let error):
-                XCTAssertNil(error)
-            case .success(let credentials):
-                XCTAssertNotNil(credentials)
-            }
-            expectation.fulfill()
+
+        do {
+            let session = URLSession.mockSession
+            let decodableObject = try await session.object(for: endpointRegister)
+            XCTAssertNil(decodableObject)
+        } catch {
+            XCTAssertNotNil(error)
+            XCTAssertNotNil(error is HTTPError)
         }
-
-        dataTask.resume()
-
-        wait(for: [expectation], timeout: 5.0)
-    }
-
-    func testDecodableError() async throws {
-        MockURLProtocol.error = URLError(.unknown)
-        MockURLProtocol.requestHandler = nil
-
-        let expectation = expectation(description: "credentials")
-        let session = URLSession.mockSession
-        let dataTask = session.dataTask(for: endpointCredentials) { result in
-            switch result {
-            case .failure(let error):
-                XCTAssertNotNil(error)
-                XCTAssertNotNil(error as? URLError)
-                XCTAssertEqual((error as! URLError).code, .unknown)
-            case .success(let credentials):
-                XCTAssertNil(credentials)
-            }
-            expectation.fulfill()
-        }
-
-        dataTask.resume()
-
-        wait(for: [expectation], timeout: 5.0)
     }
 
     // MARK: - Data
@@ -164,7 +165,7 @@ final class URLSessionEndpointTests: XCTestCase {
         MockURLProtocol.error = nil
         MockURLProtocol.requestHandler = { [unowned self] request in
             let response = HTTPURLResponse(
-                url: URL(string: "https://payment.snabble.io/apps/credentials")!,
+                url: URL(string: "https://payment.snabble.io/apps/register")!,
                 statusCode: 200,
                 httpVersion: nil,
                 headerFields: [:]
@@ -172,9 +173,9 @@ final class URLSessionEndpointTests: XCTestCase {
             return (response, resourceData)
         }
 
-        let endpoint: Endpoint<Data> = .init(path: "/apps/credentials", method: .get(nil))
+        let endpoint: Endpoint<Data> = .init(path: "/apps/register", method: .get(nil))
 
-        let expectation = expectation(description: "credentials")
+        let expectation = expectation(description: "register")
         let session = URLSession.mockSession
         session.publisher(for: endpoint)
             .sink { completion in
@@ -197,7 +198,7 @@ final class URLSessionEndpointTests: XCTestCase {
         MockURLProtocol.error = URLError(.unknown)
         MockURLProtocol.requestHandler = nil
 
-        let expectation = expectation(description: "credentials")
+        let expectation = expectation(description: "register")
         let session = URLSession.mockSession
         session.publisher(for: endpointData)
             .sink { completion in
@@ -206,7 +207,6 @@ final class URLSessionEndpointTests: XCTestCase {
                     XCTAssertTrue(true)
                 case .failure(let error):
                     XCTAssertNotNil(error)
-                    XCTAssertEqual(error.code, .unknown)
                 }
                 expectation.fulfill()
             } receiveValue: { data in
@@ -217,11 +217,43 @@ final class URLSessionEndpointTests: XCTestCase {
         wait(for: [expectation], timeout: 5.0)
     }
 
+    func testDataCombineInvalidResponse() async throws {
+        MockURLProtocol.error = nil
+        MockURLProtocol.requestHandler = { [unowned self] request in
+            let response = HTTPURLResponse(
+                url: URL(string: "https://payment.snabble.io/apps/register")!,
+                statusCode: 400,
+                httpVersion: nil,
+                headerFields: ["Content-Type": "application/json"]
+            )!
+            return (response, resourceData)
+        }
+
+        let expectation = expectation(description: "register")
+        let session = URLSession.mockSession
+        session.publisher(for: endpointData)
+            .sink { completion in
+                switch completion {
+                case .finished:
+                    XCTAssertTrue(true)
+                case .failure(let error):
+                    XCTAssertNotNil(error)
+                    XCTAssertTrue(error is HTTPError)
+                }
+                expectation.fulfill()
+            } receiveValue: { credentials in
+                XCTAssertNil(credentials)
+            }
+            .store(in: &cancellables)
+
+        wait(for: [expectation], timeout: 5.0)
+    }
+
     func testDataAsync() async throws {
         MockURLProtocol.error = nil
         MockURLProtocol.requestHandler = { [unowned self] request in
             let response = HTTPURLResponse(
-                url: URL(string: "https://payment.snabble.io/apps/credentials")!,
+                url: URL(string: "https://payment.snabble.io/apps/register")!,
                 statusCode: 200,
                 httpVersion: nil,
                 headerFields: [:]
@@ -231,7 +263,6 @@ final class URLSessionEndpointTests: XCTestCase {
 
         let session = URLSession.mockSession
         let response = try await session.data(for: endpointData)
-
         XCTAssertNotNil(response)
     }
 
@@ -245,61 +276,29 @@ final class URLSessionEndpointTests: XCTestCase {
             XCTAssertNil(decodableObject)
         } catch {
             XCTAssertNotNil(error)
-            XCTAssertNotNil(error as? URLError)
-            XCTAssertEqual((error as! URLError).code, .unknown)
+            XCTAssertNotNil(error is URLError)
         }
     }
 
-    func testData() async throws {
+    func testDataAsyncInvalidResponse() async throws {
         MockURLProtocol.error = nil
         MockURLProtocol.requestHandler = { [unowned self] request in
             let response = HTTPURLResponse(
-                url: URL(string: "https://payment.snabble.io/apps/credentials")!,
-                statusCode: 200,
+                url: URL(string: "https://payment.snabble.io/apps/register")!,
+                statusCode: 400,
                 httpVersion: nil,
-                headerFields: [:]
+                headerFields: ["Content-Type": "application/json"]
             )!
             return (response, resourceData)
         }
 
-        let session = URLSession.mockSession
-        let expectation = expectation(description: "credentials")
-        let dataTask = session.dataTask(for: endpointData) { result in
-            switch result {
-            case .failure(let error):
-                XCTAssertNil(error)
-            case .success(let response):
-                XCTAssertNotNil(response)
-            }
-            expectation.fulfill()
+        do {
+            let session = URLSession.mockSession
+            let decodableObject = try await session.object(for: endpointData)
+            XCTAssertNil(decodableObject)
+        } catch {
+            XCTAssertNotNil(error)
+            XCTAssertNotNil(error is HTTPError)
         }
-
-        dataTask.resume()
-
-        wait(for: [expectation], timeout: 5.0)
     }
-
-    func testDataError() async throws {
-        MockURLProtocol.error = URLError(.unknown)
-        MockURLProtocol.requestHandler = nil
-
-        let session = URLSession.mockSession
-        let expectation = expectation(description: "credentials")
-        let dataTask = session.dataTask(for: endpointData) { result in
-            switch result {
-            case .failure(let error):
-                XCTAssertNotNil(error)
-                XCTAssertNotNil(error as? URLError)
-                XCTAssertEqual((error as! URLError).code, .unknown)
-            case .success(let credentials):
-                XCTAssertNil(credentials)
-            }
-            expectation.fulfill()
-        }
-
-        dataTask.resume()
-
-        wait(for: [expectation], timeout: 5.0)
-    }
-
 }